requirements.txt
opencv-python
imutils
scikit-image
plantcv
img2vec_pytorch
pillow
matplotlib
seaborn
